Update Atk_Aoe_Manager.xml
Now works with skills and you can even set your own keyword for it to use instead of the skill name for those like Kobold spray that uses kspray to work. Even if the skill is on the list as Kobold spray when you use it it will be kspray. This is so it can still autoupdate from the list like the spells do but still work correctly when used.