Update Potion_tracker.xml
Added a new ptw status to show all your current settings. Fixed both the ptw kw list and ptw list commands to be much easier to read. Fixed the helpfile to show the new command and overhauled it to look better and categorized it. Added version info to status and helpfile also.