Update Potion_tracker.xml
Went ahead and added ptw usage on/off to be able to cut the usage messages off but check them out first they now look much better and align with the way the window does the color thresholds so they change based on those as you use potions also. Bumped version for this neat addition.